feat: 뉴스 서비스에 회원 정보 및 북마크 여부 반영 (#11)
`/news/controller/NewsController`
- 일반 뉴스 목록 조회시 회원 인증 여부 검사 및 미인증 회원 허용
- 뉴스 상세 조회시 회원 인증 여부 검사 및 미인증 회원 허용
- 뉴스 상세 조회시 댓글 속성 미포함
- 뉴스 생성시 회원 인증 검사 및 예외 처리
- 뉴스 업데이트시 회원 인증 검사 및 예외 처리
- 뉴스 삭제시 회원 인증 및 관리자 여부 검사와 예외 처리

+++ backend/src/main/java/com/tamnara/backend/news/controller/NewsController.java
@@ -5,9 +5,12 @@ import com.tamnara.backend.news.dto.NewsCardDTO;
 import com.tamnara.backend.news.dto.request.NewsCreateRequest;
 import com.tamnara.backend.news.dto.response.NewsDetailResponse;
 import com.tamnara.backend.news.service.NewsService;
+import com.tamnara.backend.user.domain.Role;
+import com.tamnara.backend.user.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -54,9 +57,15 @@ public class NewsController {
 
     @GetMapping
     public ResponseEntity<?> findNormalNews(
+            @AuthenticationPrincipal UserDetailsImpl userDetails,
             @RequestParam(required = false) String category,
             @RequestParam(defaultValue = "0") Integer offset) {
         try {
+            Long userId = null;
+            if (!(userDetails == null || userDetails.getUser() == null)) {
+                userId = userDetails.getUser().getId();
+            }
+
             int pageNum = offset / PAGE_SIZE;
             int nextOffset = (pageNum + 1) * PAGE_SIZE;
 
@@ -66,9 +75,9 @@ public class NewsController {
                     throw new IllegalArgumentException("추가 요청일 경우 offset은 0이어야 합니다.");
                 }
 
-                List<NewsCardDTO> newsCards = newsService.getNewsCardPage(null, false, category, pageNum, PAGE_SIZE);
+                List<NewsCardDTO> newsCards = newsService.getNewsCardPage(userId, false, category, pageNum, PAGE_SIZE);
 
-                boolean hasNext = !newsService.getNewsCardPage(null, false, category, pageNum + 1, PAGE_SIZE).isEmpty();
+                boolean hasNext = !newsService.getNewsCardPage(userId, false, category, pageNum + 1, PAGE_SIZE).isEmpty();
 
                 Map<String, Object> data = Map.of(
                         category, newsCards,
@@ -87,7 +96,7 @@ public class NewsController {
                     throw new IllegalArgumentException("최초 요청일 경우 offset은 0보다 큰 값이어야 합니다.");
                 }
 
-                Map<String, List<NewsCardDTO>> newsCardsMap = newsService.getNormalNewsCardPages(null, false, pageNum, PAGE_SIZE);
+                Map<String, List<NewsCardDTO>> newsCardsMap = newsService.getNormalNewsCardPages(userId, false, pageNum, PAGE_SIZE);
                 Map<String, Object> data = new HashMap<>();
 
                 for (Map.Entry<String, List<NewsCardDTO>> entry : newsCardsMap.entrySet()) {
@@ -126,15 +135,18 @@ public class NewsController {
     }
 
     @GetMapping("/{newsId}")
-    public ResponseEntity<?> findNewsDetail(@PathVariable("newsId") Long newsId) {
+    public ResponseEntity<?> findNewsDetail(@PathVariable("newsId") Long newsId,
+                                            @AuthenticationPrincipal UserDetailsImpl userDetails
+    ) {
         try {
-            NewsDetailResponse res = newsService.getNewsDetail(newsId, null);
+            Long userId = null;
+            if (!(userDetails == null || userDetails.getUser() == null)) {
+                userId = userDetails.getUser().getId();
+            }
 
-            Map<String, Object> data = Map.of(
-                    "news", res,
-                    "comment", ""
-            );
+            NewsDetailResponse res = newsService.getNewsDetail(newsId, userId);
 
+            Map<String, Object> data = Map.of("news", res);
             return ResponseEntity.status(HttpStatus.OK).body(Map.of(
                     "success", true,
                     "message", "요청하신 데이터를 성공적으로 불러왔습니다.",
@@ -151,9 +163,16 @@ public class NewsController {
     }
 
     @PostMapping
-    public ResponseEntity<?> createNews(@ModelAttribute NewsCreateRequest req) {
+    public ResponseEntity<?> createNews(@AuthenticationPrincipal UserDetailsImpl userDetails,
+                                        @ModelAttribute NewsCreateRequest req
+    ) {
         try {
-            NewsDetailResponse res = newsService.save(null, false, req);
+            if (userDetails == null || userDetails.getUser() == null) {
+                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "인증되지 않은 회원입니다.");
+            }
+
+            Long userId = userDetails.getUser().getId();
+            NewsDetailResponse res = newsService.save(userId, false, req);
             return ResponseEntity.status(HttpStatus.CREATED).body(Map.of(
                     "success", true,
                     "message", "데이터가 성공적으로 생성되었습니다.",
@@ -170,9 +189,16 @@ public class NewsController {
     }
 
     @PatchMapping("/{newsId}")
-    public ResponseEntity<?> updateNews(@PathVariable Long newsId) {
+    public ResponseEntity<?> updateNews(@PathVariable Long newsId,
+                                        @AuthenticationPrincipal UserDetailsImpl userDetails
+    ) {
         try {
-            NewsDetailResponse res = newsService.update(newsId, null);
+            if (userDetails == null || userDetails.getUser() == null) {
+                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "인증되지 않은 회원입니다.");
+            }
+
+            Long userId = userDetails.getUser().getId();
+            NewsDetailResponse res = newsService.update(newsId, userId);
             return ResponseEntity.status(HttpStatus.OK).body(Map.of(
                     "success", true,
                     "message", "데이터가 성공적으로 업데이트되었습니다.",
@@ -189,8 +215,18 @@ public class NewsController {
     }
 
     @DeleteMapping("/{newsId}")
-    public ResponseEntity<?> deleteNews(@PathVariable Long newsId) {
+    public ResponseEntity<?> deleteNews(@PathVariable Long newsId,
+                                        @AuthenticationPrincipal UserDetailsImpl userDetails
+    ) {
         try {
+            if (userDetails == null || userDetails.getUser() == null) {
+                throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "인증되지 않은 회원입니다.");
+            }
+
+            if (userDetails.getUser().getRole() != Role.ADMIN) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "뉴스를 삭제할 권한이 없습니다.");
+            }
+
             Long resNewsId = newsService.delete(newsId, null);
 
             Map<String, Object> data = Map.of("newsId", resNewsId);
